requirements.txt
Click==8.0.4
Flask==2.0.1
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
requests
seaborn
matplotlib
pandas
pytest-flask